items now spawn in chests, chests can be searched

+++ src/components/dungeon_level.rs
@@ -7,5 +7,5 @@ use specs::{
 
 #[derive(Component, ConvertSaveload, Clone, Debug)]
 pub struct DungeonLevel {
-  pub level: i32,
+  pub level: u8,
 }
